crypto/tls: Fix a bug in the ECH extension
To compute ClientHelloOuterAAD, the server unmarshals ClientHelloOuter,
removes the ECH extension, then marshal the result the same way the
client does in this package. As a result, the server may re-order and
remove unrecognized extensions in the processed. This changes the
computation of ClientHelloOuterAAD to operate on the serialized
ClientHelloOuter.