Improve CIRCL update script
Removes cf-copy-from-circle.py and replaces it with cf-update-circl.py.
Instead of copying everything and removing unused packages,
cf-update-circl.py defines a manifest of packages that are needed for
the standard library and copies them. This reduces the risk of changes
to CIRCL breaking the build of the Go standard library.